rearrange tropo_delay imports for raider-base
